bot/requirements.txt
motor>=2.3
aioredis>=1.3
aiohttp>=3.7
discord.py==1.5.1